+++ packages/terra-application-name/src/terra-dev-site/doc/example/ApplicationHeaderNameStandard.jsx
@@ -7,22 +7,23 @@ import Placeholder from 'terra-application-name/lib/terra-dev-site/doc/common/Pl
 import ApplicationHeaderName from 'terra-application-name/lib/ApplicationHeaderName';
 /* eslint-enable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 import demoColors from '../../test/application-name/demoStyles.scss';
+import styles from './ApplicationHeaderNameStandard.scss';
 
 const cx = classNames.bind(demoColors);
-const demosize = { height: '50px', width: '100%' };
+const cy = classNames.bind(styles);
 const ApplicationHeaderNameStandard = () => (
   <ApplicationHeaderLayout
-    style={demosize}
+    className={cy('demo-size')}
     logo={(
       <ApplicationHeaderName
         title="App-Name"
-        accessory={<Image alt="Terra Logo" variant="rounded" src="https://github.com/cerner/terra-framework/raw/master/terra.png" height="26px" width="26px" isFluid />}
+        accessory={<Image alt="Terra Logo" variant="rounded" src="https://github.com/cerner/terra-framework/raw/master/terra.png" className={cx('demo-image-container')} isFluid />}
         className={cx(['demo-background-color'])}
       />
     )}
-    extensions={<Placeholder text="Extensions" width="100px" />}
-    navigation={<Placeholder text="Content" />}
-    utilities={<Placeholder text="Utilities" width="100px" />}
+    extensions={<Placeholder text="Extensions" type="header-extensions" />}
+    navigation={<Placeholder text="Content" type="default" />}
+    utilities={<Placeholder text="Utiltities" type="utiltities" />}
   />
 );