+++ packages/terra-modal-manager/src/terra-dev-site/test/modal-manager/DisclosureIntegrationComponent.jsx
@@ -5,6 +5,10 @@ import SelectField from 'terra-form-select';
 import {
   availableDisclosureHeights, availableDisclosureWidths, withDisclosureManager, disclosureManagerShape,
 } from 'terra-disclosure-manager';
+import classNames from 'classnames/bind';
+import styles from './DisclosureComponentCommon.scss';
+
+const cx = classNames.bind(styles);
 
 const HEIGHT_KEYS = Object.keys(availableDisclosureHeights);
 const WIDTH_KEYS = Object.keys(availableDisclosureWidths);
@@ -117,7 +121,7 @@ class DisclosureComponent extends React.Component {
     const { disclosureManager, identifier } = this.props;
 
     return (
-      <ContentContainer id={identifier} className="nested-component" fill header={<h2 style={{ margin: '0', borderBottom: '1px solid black' }}>Content Component</h2>}>
+      <ContentContainer id={identifier} className="nested-component" fill header={<h2 className={cx('header')}>Content Component</h2>}>
         <p>
           id:
           {' '}
@@ -130,12 +134,12 @@ class DisclosureComponent extends React.Component {
         <button type="button" className="disclose-large" onClick={this.disclose('large')}>Disclose Large</button>
         <button type="button" className="disclose-huge" onClick={this.disclose('huge')}>Disclose Huge</button>
         <button type="button" className="disclose-fullscreen" onClick={this.disclose('fullscreen')}>Disclose Fullscreen</button>
-        <div style={{ padding: '0.7rem' }}>
+        <div className={cx('content-form')}>
           {this.renderForm()}
           {this.renderFormButton()}
           <br />
           <br />
-          <SelectField label="T-shirt size" placeholder="Select a size" selectId="tshirt-size-field" required style={{ maxWidth: '300px' }}>
+          <SelectField label="T-shirt size" placeholder="Select a size" selectId="tshirt-size-field" required className={cx('select-field')}>
             <SelectField.Option value="Red" display="Red" />
             <SelectField.Option value="Orange" display="Orange" />
             <SelectField.Option value="Yellow" display="Yellow" />