+++ packages/terra-application-utility/tests/wdio/application-header-utility-spec.js
@@ -1,26 +1,20 @@
-const viewports = Terra.viewports('tiny', 'medium');
+Terra.describeViewports('ApplicationHeaderUtility', ['tiny', 'medium'], () => {
+  before(() => {
+    browser.url('/#/raw/tests/terra-application-utility/application-utility/default-application-header-utility');
+    browser.waitForVisible('#default');
+  });
 
-viewports.forEach((viewport) => {
-  describe(`ApplicationHeaderUtility - ${viewport}`, () => {
-    before(() => browser.setViewportSize(viewport));
+  describe('Displays a default application header utility', () => {
+    Terra.it.validatesElement({ selector: '#default' });
+  });
 
-    beforeEach(() => {
-      browser.url('/#/raw/tests/terra-application-utility/application-utility/default-application-header-utility');
-      browser.waitForVisible('#default');
-    });
+  describe('Hover-application header utility', () => {
+    before(() => { browser.moveToObject('#default'); });
+    Terra.it.validatesElement({ selector: '#default' });
+  });
 
-    describe('Displays a default application header utility', () => {
-      Terra.it.validatesElement({ selector: '#default' });
-    });
-
-    describe('Hover-application header utility', () => {
-      beforeEach(() => { browser.moveToObject('#default'); });
-      Terra.it.validatesElement({ selector: '#default' });
-    });
-
-    describe('Focus-application header utility', () => {
-      beforeEach(() => { browser.keys('Tab'); });
-      Terra.it.validatesElement({ selector: '#default' });
-    });
+  describe('Focus-application header utility', () => {
+    before(() => { browser.keys('Tab'); });
+    Terra.it.validatesElement({ selector: '#default' });
   });
 });