+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 * Added check to only load inert polyfill if it is not already defined on the Element prototype
 
 ### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
 * Update wdio spec to use Terra.describeViewports
 
 3.2.0 - (June 12, 2019)