+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -3,11 +3,15 @@ Changelog
 
 Unreleased
 ----------
+
+3.3.0 - (June 19, 2019)
+------------------
 ### Added
 * Added check to only load inert polyfill if it is not already defined on the Element prototype
 
 ### Changed
 * Removed all inline styles from Dev-site components and implemented the same styles using external css.
+* Update wdio spec to use Terra.describeViewports
 
 3.2.0 - (June 12, 2019)
 ------------------